requirements.txt
psycopg2==2.9.10
python-dotenv==1.0.1
tabulate==0.9.0
